refactor: move the sign components into modules/signs
The 3D half first: Signs and AutoSignage, with their tests, into the shape
modules/initiative already uses. Both drew the same kind of object but sat loose
in components/ with the rest of the app.

Moved with git mv so the history follows, and every caller now goes through the
module's index rather than reaching for a file — App, AdminPanel and useMapData
were the only three.

Two things share the module because they draw the same kind of object, not
because they are one feature: signs are placed by hand and stored in the
database, AutoSignage is generated from where the buildings are and stored
nowhere. The index says so.

Nothing under backend/ is touched, on this branch or in the tree, and the 24
signs API tests still pass — which is what keeps existing signs safe. The editor
UI is still inside AdminPanel and moves next.
