+++ frontend/src/modules/signs/__tests__/SignEditingSelection.test.tsx
